feat: add styles for the component options bar (no component selected)
Develop the styles for the component options bar when any component is
selected from the canvas. Fix bug when doing drag and drop from several
components from the components bar (the component being moved would be the
the one to the right of the bar, not the selected one). Minor fixes on
the home and components bar styles.