fix: improve styles of the components bar, canvas and styles bar
Adjust the styles from the components bar, canvas and styles bar so it
can fit better into the screen. The components and styles bars both have
fixed width in order to keep the same size in case the window resizes or
not. The canvas, on the other hand, have a resizable width. All of these
in order to have a better UX and achieve a similar behavior as of
Figma.