feat: generate html and css files
The html and the css files are generated when the Generate button is
clicked. A native file picker is shown so the user can selected the
desired location for the page file folder. This functionality works with
an invoke call, and the response is a string. A popup panel indicating
the result will be developed later.