feat: improve generating page files
Improved the functionality for generating the page files. The default
styles are now added to the css rule of the components. A UUID is
generated for each component and it's set as its id so the selection in
the css file is done with the id to avoid style mismatching.